Remove violation tree
